oikoitea/requirements.txt
boto==2.48.0
boto3==1.5.8
botocore==1.8.22
bz2file==0.98
certifi==2017.11.5
chardet==3.0.4
Django==1.11.5
django-adminlte2==0.1.6
django-crispy-forms==1.6.1
django-el-pagination==3.1.0
django-formset-js==0.5.0
django-jquery-js==3.1.1
docutils==0.14
futures==3.2.0
gensim==3.2.0
idna==2.6
jmespath==0.9.3
mysqlclient==1.3.12
numpy==1.13.3
olefile==0.44
Pillow==4.2.1
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
s3transfer==0.1.12
scipy==1.0.0
six==1.11.0
smart-open==1.5.6
urllib3==1.22
